Missing error message during polarion test export (#4181)
During export to Polarion there was no error message when an fmf tree is
not part of any git repository. Adding this message and including
functionality related to ignore-git-validation, from now it will be
possible to export with local repository for both nitrate and polarion
test export, intended for testing with dry mode but possible to run with
actual export as well.

Fixes #4175 

Co-authored-by: Petr Matyas <pmatyas@redhat.com>