Extra pidfile handling into its own context class (#4236)
A followup for #3857 and #4232, this time liberating test pidfile from
the internal `execute` plugin. The goal is the same, reuse various
existing pieces in prepare and finish, to reduce maintenance and deliver
the same behavior.; to support `tmt-reboot`, even `prepare/shell` will
need to maintain a pidfile for scripts its running.

Related to #902.