Stop filtering src RPMs in Repository Provider, add unit tests (#4207)
This PR bundles several improvements for the `repository` provider.

The main functional change removes the filtering logic that was
previously (and incorrectly) excluding `src` RPMs from being processed
as artifacts.

Additionally, this PR introduces unit tests for the `repository`
provider to improve code coverage and ensure reliability. It also
includes a minor correction to the provider's name.