Drop the hatch cache usage (#4281)
Every now and then I see the github action PR failing mostly seeming to
be related to the `tmt` version that is being used by sphinx, which
might be using a cached version. Sphinx does a lot of magic around the
import, and I am not even comfortable keeping it in a non-src-layout
format, but we should discuss that in
https://github.com/teemtee/tmt/issues/3118.

For this change at least we should make sure we always have a fresh
hatch environment, and all relevant wheel caches should be handled by
`setup-python` instead.

Signed-off-by: Cristian Le <git@lecris.dev>