Implement the `TMT_PLAN_SOURCE_SCRIPT` hook (#4045)
Generate a file for each plan that is sourced at the beginning of each:
- prepare.shell step
- test invocation

These are explicitly not invoked on all internal guest execute in order
to minimize side-effects, and instead these are meant to be used by the
user for their test environment. This file is thus exposed as
`TMT_PLAN_SOURCE_SCRIPT` and it is a no-op until they populate it via
another prepare step or test invocation.

Closes #3981

---------

Signed-off-by: Cristian Le <git@lecris.dev>