Make sure the required dependencies for bootc are present (#4403)
A proper fix is to do a similar `import` test that we do in the other
plugins, but we do not seem to have a central place to do so, so punting
that for afterwards (https://github.com/teemtee/tmt/issues/4404). Half
of the issue reported will not be addressed, i.e. having a good error
message for installing the appropriate `tmt[provision-bootc]` but I
think we should postpone this and find a general solution across the
board.

For now here we just make sure that `podman-machine` is always present
according to the general usage

https://github.com/teemtee/tmt/blob/20bf98135f93ddfd84e2010aeb7118bcb7671234/tmt/steps/provision/bootc.py#L348-L352

I have also patched up a seemingly missing relation in `pyproject.toml`

---

Closes #4381

---------

Signed-off-by: Cristian Le <git@lecris.dev>