Make the sphinx autodoc documents evaluate last (#4286)
This is quite a nasty hidden issue. Because both sphinx and tmt are
using annotations, if sphinx imports tmt first, it will affect the code
when we actually want to evaluate the tmt modules.

I could not find a way around it and until upstream offers some
interface https://github.com/sphinx-doc/sphinx/issues/14010, this is the
best workaround that we can do.

Depends on: #3977

---------

Signed-off-by: Cristian Le <git@lecris.dev>