Refactor `shell` plugin to produce results (#4307)
* Well, it really, really should. Users should not dive into `log.txt`.
* It's also needed, to support `tmt-reboot` in `prepare` and `finish` -
it turns out reboot support is much easier to implement when different
outcomes of a script are handled separately. That is, success, failure,
error, all these can be then extended with reboot support similar to the
one in `execute`, i.e. with the use of context.

So, the patch changes `shell` plugin which now produces results and
logs. Next patches will extend some of the branches with reboot support
which will fit more nicely. Implementing reboot support first, then
adding support for results would be terrible and ugly.

Pull Request Checklist

* [x] implement the feature