Add `uv.lock`/`pylock.toml` files (#4368)
This is the initial introduction of the uv.lock/pylock.toml files.
Ideally it would be nice to use only the more agnostic `pylock.toml` but
as far as I was told[^1], this is not yet possible, so for now we have
to carry both files. Currently only some places are connected to the
lock files:
- RTD builders
- Github actions
- Renovate (should be built-in)

Out of scope for now:
- Documenting developer's setting of uv and hooking it to the makefile
#4369
- Covering the various `uv install` in the `pip install` tests #4370
- Moving the hatch environments to something backend independent #4371
- Running any sort of tests using the uv lock-ed environments #4370
- Use uv for some pre-commits (mypy, pyright) #4372

Relates to #4336

[^1]:
https://github.com/astral-sh/uv/issues/12584#issuecomment-3571748311

---------

Signed-off-by: Cristian Le <git@lecris.dev>