Support FIPS prepare feature on RHEL-7 (#4395)
Notice that this requires python-3.9 or newer to be present on guest
that is not present by default on RHEL-7. Hence, a provisioning or
previous prepare steps must handle this (e.g. testing-farm is doing this
correctly).

Changes:

1. *Disable prelink* - following**
https://access.redhat.com/solutions/137833#initial (Initial Common
Steps) I added prelink disabling. Notice that this (by default) only
affects RHEL-7 since RHEL-8 comes with prelink disabled already and
RHEL-9 and 10 (as well as their Centos Stream variants) do not have
prelink anymore. However, I did not limit this step to RHEL-7 just to be
100% sure there is either no prelink or that it is disabled. Prelinking
would make FIPS 140 power-on-self-test fail for all cryptographic
backends and make system unusable.

2. *Modify bootloader* - this step was previously only done on RHEL-10
but it was essentially happening also on RHEL-8 and RHEL-9 via "Enable
FIPS mode" step (`fips-mode-setup --enable`). Now, this exact step is
also needed for RHEL-7 and hence for clarify it makes sense to do the
bootloader update the same way for all releases while excluding the
change from "Enable FIPS mode" (see below).

3. *Enable FIPS mode* - `fips-mode-setup --enable` done on RHEL-8 and
RHEL-9 updates bootloader settings (adding `fips=1` parameter and
`boot=...` parameter if it is needed). Since we need do to bootloader
update for RHEL-7 and RHEL-10 explicitly by "Modify bootloader" steps,
why not to do it for all releases for simplicity? In order not to do it
twice for RHEL-8 and RHEL-9 `--no-bootcfg` is added for
`fips-mode-setup` to tell the tool that we don't want it to do the
bootloader modifications.

Pull Request Checklist

* [x] implement the feature

Signed-off-by: Ondrej Moris <omoris@redhat.com>