Move test check invocation under the umbrella of test invocation
After moving the environment, they can be simplified a lot, fixing the
use of the right logger in the process. Since the test invocation class
is the owner of "invoke me a test, please" task, and invoking test
checks is part of this task (a test result without check results is
incomplete), it's the natural place to own this piece of code.